Switch to using dw_main_sleep() instead of msleep() which isn't GUI aware. Switch to using long long to store file sizes to avoid 2GB file limit. Switch to using stat64() on supported platforms to avoid 2GB file limit. Only tested on Mac... may require fixes for other platforms.

+++ src/site.h
@@ -3,7 +3,7 @@
 typedef struct _directory {
 	char *entry;
 	struct _directory *next;
-	ULONG size;
+	unsigned long long size;
 	CTIME time;
 	CDATE date;
 	int type;
@@ -15,7 +15,7 @@ typedef struct _queue {
 	/*SiteTab *dest;*/
 	char *site;
 	char *destdirectory;
-	ULONG size;
+	unsigned long long size;
 	CTIME time;
 	CDATE date;
 	int type;
